+++ src/rt/arc.rs
@@ -71,26 +71,26 @@ impl Arc {
                 last_ref_modification: None,
             });
 
-            trace!(?state, "Arc::new");
+            trace!(?state, %location, "Arc::new");
 
             Arc { state }
         })
     }
 
-    pub(crate) fn ref_inc(&self) {
-        self.branch(Action::RefInc);
+    pub(crate) fn ref_inc(&self, location: Location) {
+        self.branch(Action::RefInc, location);
 
         rt::execution(|execution| {
             let state = self.state.get_mut(&mut execution.objects);
             state.ref_cnt = state.ref_cnt.checked_add(1).expect("overflow");
 
-            trace!(state = ?self.state, ref_cnt = ?state.ref_cnt, "Arc::ref_inc");
+            trace!(state = ?self.state, ref_cnt = ?state.ref_cnt, %location, "Arc::ref_inc");
         })
     }
 
     /// Validate a `get_mut` call
-    pub(crate) fn get_mut(&self) -> bool {
-        self.branch(Action::RefDec);
+    pub(crate) fn get_mut(&self, location: Location) -> bool {
+        self.branch(Action::RefDec, location);
 
         rt::execution(|execution| {
             let state = self.state.get_mut(&mut execution.objects);
@@ -102,15 +102,15 @@ impl Arc {
 
             let is_only_ref = state.ref_cnt == 1;
 
-            trace!(state = ?self.state, ?is_only_ref, "Arc::get_mut");
+            trace!(state = ?self.state, ?is_only_ref, %location, "Arc::get_mut");
 
             is_only_ref
         })
     }
 
     /// Returns true if the memory should be dropped.
-    pub(crate) fn ref_dec(&self) -> bool {
-        self.branch(Action::RefDec);
+    pub(crate) fn ref_dec(&self, location: Location) -> bool {
+        self.branch(Action::RefDec, location);
 
         rt::execution(|execution| {
             let state = self.state.get_mut(&mut execution.objects);
@@ -120,7 +120,7 @@ impl Arc {
             // Decrement the ref count
             state.ref_cnt -= 1;
 
-            trace!(state = ?self.state, ref_cnt = ?state.ref_cnt, "Arc::ref_dec");
+            trace!(state = ?self.state, ref_cnt = ?state.ref_cnt, %location, "Arc::ref_dec");
 
             // Synchronize the threads.
             state
@@ -154,9 +154,9 @@ impl Arc {
         })
     }
 
-    fn branch(&self, action: Action) {
+    fn branch(&self, action: Action, location: Location) {
         let r = self.state;
-        r.branch_action(action);
+        r.branch_action(action, location);
         assert!(
             r.ref_eq(self.state),
             "Internal state mutated during branch. This is \
@@ -167,12 +167,15 @@ impl Arc {
 }
 
 impl State {
-    pub(super) fn check_for_leaks(&self) {
+    pub(super) fn check_for_leaks(&self, index: usize) {
         if self.ref_cnt != 0 {
             if self.allocated.is_captured() {
-                panic!("Arc leaked.\n  Allocated: {}", self.allocated);
+                panic!(
+                    "Arc leaked.\n  Allocated: {}\n      Index: {}",
+                    self.allocated, index
+                );
             } else {
-                panic!("Arc leaked.");
+                panic!("Arc leaked.\n  Index: {}", index);
             }
         }
     }