+++ src/rt/thread.rs
@@ -3,6 +3,8 @@ use crate::rt::object::Operation;
 use crate::rt::vv::VersionVec;
 
 use std::{any::Any, collections::HashMap, fmt, ops};
+
+use super::Location;
 pub(crate) struct Thread {
     pub id: Id,
 
@@ -31,6 +33,9 @@ pub(crate) struct Thread {
     pub yield_count: usize,
 
     locals: LocalMap,
+
+    /// `tracing` span used to associate diagnostics with the current thread.
+    span: tracing::Span,
 }
 
 #[derive(Debug)]
@@ -49,6 +54,9 @@ pub(crate) struct Set {
     /// Sequential consistency causality. All sequentially consistent operations
     /// synchronize with this causality.
     pub seq_cst_causality: VersionVec,
+
+    /// `tracing` span used as the parent for new thread spans.
+    iteration_span: tracing::Span,
 }
 
 #[derive(Eq, PartialEq, Hash, Copy, Clone)]
@@ -67,8 +75,8 @@ impl Id {
 
 #[derive(Debug, Clone, Copy)]
 pub(crate) enum State {
-    Runnable,
-    Blocked,
+    Runnable { unparked: bool },
+    Blocked(Location),
     Yield,
     Terminated,
 }
@@ -81,10 +89,11 @@ struct LocalKeyId(usize);
 struct LocalValue(Option<Box<dyn Any>>);
 
 impl Thread {
-    fn new(id: Id) -> Thread {
+    fn new(id: Id, parent_span: &tracing::Span) -> Thread {
         Thread {
             id,
-            state: State::Runnable,
+            span: tracing::info_span!(parent: parent_span.id(), "thread", id = id.id),
+            state: State::Runnable { unparked: false },
             critical: false,
             operation: None,
             causality: VersionVec::new(),
@@ -97,32 +106,23 @@ impl Thread {
     }
 
     pub(crate) fn is_runnable(&self) -> bool {
-        match self.state {
-            State::Runnable => true,
-            _ => false,
-        }
+        matches!(self.state, State::Runnable { .. })
     }
 
     pub(crate) fn set_runnable(&mut self) {
-        self.state = State::Runnable;
+        self.state = State::Runnable { unparked: false };
     }
 
-    pub(crate) fn set_blocked(&mut self) {
-        self.state = State::Blocked;
+    pub(crate) fn set_blocked(&mut self, location: Location) {
+        self.state = State::Blocked(location);
     }
 
     pub(crate) fn is_blocked(&self) -> bool {
-        match self.state {
-            State::Blocked => true,
-            _ => false,
-        }
+        matches!(self.state, State::Blocked(..))
     }
 
     pub(crate) fn is_yield(&self) -> bool {
-        match self.state {
-            State::Yield => true,
-            _ => false,
-        }
+        matches!(self.state, State::Yield)
     }
 
     pub(crate) fn set_yield(&mut self) {
@@ -132,10 +132,7 @@ impl Thread {
     }
 
     pub(crate) fn is_terminated(&self) -> bool {
-        match self.state {
-            State::Terminated => true,
-            _ => false,
-        }
+        matches!(self.state, State::Terminated)
     }
 
     pub(crate) fn set_terminated(&mut self) {
@@ -146,7 +143,7 @@ impl Thread {
         let mut locals = Vec::with_capacity(self.locals.len());
 
         // run the Drop impls of any mock thread-locals created by this thread.
-        for (_, local) in &mut self.locals {
+        for local in self.locals.values_mut() {
             locals.push(local.0.take());
         }
 
@@ -155,9 +152,16 @@ impl Thread {
 
     pub(crate) fn unpark(&mut self, unparker: &Thread) {
         self.causality.join(&unparker.causality);
+        self.set_unparked();
+    }
 
+    /// Unpark a thread's state. If it is already runnable, store the unpark for
+    /// a future call to `park`.
+    fn set_unparked(&mut self) {
         if self.is_blocked() || self.is_yield() {
             self.set_runnable();
+        } else if self.is_runnable() {
+            self.state = State::Runnable { unparked: true }
         }
     }
 }
@@ -187,15 +191,18 @@ impl Set {
     /// The set may contain up to `max_threads` threads.
     pub(crate) fn new(execution_id: execution::Id, max_threads: usize) -> Set {
         let mut threads = Vec::with_capacity(max_threads);
-
+        // Capture the current iteration's span to be used as each thread
+        // span's parent.
+        let iteration_span = tracing::Span::current();
         // Push initial thread
-        threads.push(Thread::new(Id::new(execution_id, 0)));
+        threads.push(Thread::new(Id::new(execution_id, 0), &iteration_span));
 
         Set {
             execution_id,
             threads,
             active: Some(0),
             seq_cst_causality: VersionVec::new(),
+            iteration_span,
         }
     }
 
@@ -211,8 +218,10 @@ impl Set {
         let id = self.threads.len();
 
         // Push the thread onto the stack
-        self.threads
-            .push(Thread::new(Id::new(self.execution_id, id)));
+        self.threads.push(Thread::new(
+            Id::new(self.execution_id, id),
+            &self.iteration_span,
+        ));
 
         Id::new(self.execution_id, id)
     }
@@ -234,6 +243,15 @@ impl Set {
     }
 
     pub(crate) fn set_active(&mut self, id: Option<Id>) {
+        tracing::dispatcher::get_default(|subscriber| {
+            if let Some(span_id) = self.active().span.id() {
+                subscriber.exit(&span_id)
+            }
+
+            if let Some(span_id) = id.and_then(|id| self.threads.get(id.id)?.span.id()) {
+                subscriber.enter(&span_id);
+            }
+        });
         self.active = id.map(Id::as_usize);
     }
 
@@ -269,12 +287,16 @@ impl Set {
 
     pub(crate) fn unpark(&mut self, id: Id) {
         if id == self.active_id() {
+            // The thread is unparking itself. We don't have to join its
+            // causality with the unparker's causality in this case, since the
+            // thread *is* the unparker. Just unpark its state.
+            self.active_mut().set_unparked();
             return;
         }
 
         // Synchronize memory
         let (active, th) = self.active2_mut(id);
-        th.unpark(&active);
+        th.unpark(active);
     }
 
     /// Insert a point of sequential consistency
@@ -310,15 +332,17 @@ impl Set {
     }
 
     pub(crate) fn clear(&mut self, execution_id: execution::Id) {
+        self.iteration_span = tracing::Span::current();
         self.threads.clear();
-        self.threads.push(Thread::new(Id::new(execution_id, 0)));
+        self.threads
+            .push(Thread::new(Id::new(execution_id, 0), &self.iteration_span));
 
         self.execution_id = execution_id;
         self.active = Some(0);
         self.seq_cst_causality = VersionVec::new();
     }
 
-    pub(crate) fn iter<'a>(&'a self) -> impl ExactSizeIterator<Item = (Id, &'a Thread)> + 'a {
+    pub(crate) fn iter(&self) -> impl ExactSizeIterator<Item = (Id, &Thread)> + '_ {
         let execution_id = self.execution_id;
         self.threads
             .iter()
@@ -326,9 +350,7 @@ impl Set {
             .map(move |(id, thread)| (Id::new(execution_id, id), thread))
     }
 
-    pub(crate) fn iter_mut<'a>(
-        &'a mut self,
-    ) -> impl ExactSizeIterator<Item = (Id, &'a mut Thread)> {
+    pub(crate) fn iter_mut(&mut self) -> impl ExactSizeIterator<Item = (Id, &mut Thread)> + '_ {
         let execution_id = self.execution_id;
         self.threads
             .iter_mut()