+++ src/sync/atomic/bool.rs
@@ -3,6 +3,9 @@ use super::Atomic;
 use std::sync::atomic::Ordering;
 
 /// Mock implementation of `std::sync::atomic::AtomicBool`.
+///
+/// NOTE: Unlike `std::sync::atomic::AtomicBool`, this type has a different
+/// in-memory representation than `bool`.
 #[derive(Debug)]
 pub struct AtomicBool(Atomic<bool>);
 
@@ -14,6 +17,13 @@ impl AtomicBool {
     }
 
     /// Load the value without any synchronization.
+    ///
+    /// # Safety
+    ///
+    /// An unsynchronized atomic load technically always has undefined behavior.
+    /// However, if the atomic value is not currently visible by other threads,
+    /// this *should* always be equivalent to a non-atomic load of an un-shared
+    /// `bool` value.
     #[track_caller]
     pub unsafe fn unsync_load(&self) -> bool {
         self.0.unsync_load()