+++ src/sync/notify.rs
@@ -27,17 +27,25 @@ impl Notify {
     }
 
     /// Notify the waiter
+    #[track_caller]
     pub fn notify(&self) {
-        self.object.notify();
+        self.object.notify(location!());
     }
 
     /// Wait for a notification
+    #[track_caller]
     pub fn wait(&self) {
         self.waiting
             .compare_exchange(false, true, SeqCst, SeqCst)
             .expect("only a single thread may wait on `Notify`");
 
-        self.object.wait();
+        self.object.wait(location!());
         self.waiting.store(false, SeqCst);
+    }
+}
+
+impl Default for Notify {
+    fn default() -> Self {
+        Self::new()
     }
 }