+++ src/thread.rs
@@ -2,7 +2,7 @@
 
 pub use crate::rt::thread::AccessError;
 pub use crate::rt::yield_now;
-use crate::rt::{self, Execution};
+use crate::rt::{self, Execution, Location};
 
 pub use std::thread::panicking;
 
@@ -34,7 +34,20 @@ impl Thread {
 
     /// Returns the (optional) name of this thread
     pub fn name(&self) -> Option<&str> {
-        self.name.as_ref().map(|s| s.as_str())
+        self.name.as_deref()
+    }
+
+    /// Mock implementation of [`std::thread::unpark`].
+    ///
+    /// Atomically makes the handle's token available if it is not already.
+    ///
+    /// Every thread is equipped with some basic low-level blocking support, via
+    /// the [`park`][park] function and the `unpark()` method. These can be
+    /// used as a more CPU-efficient implementation of a spinlock.
+    ///
+    /// See the [park documentation][park] for more details.
+    pub fn unpark(&self) {
+        rt::execution(|execution| execution.threads.unpark(self.id.id));
     }
 }
 
@@ -107,16 +120,28 @@ pub fn current() -> Thread {
 ///
 /// Note that you may only have [`MAX_THREADS`](crate::MAX_THREADS) threads in a given loom tests
 /// _including_ the main thread.
+#[track_caller]
 pub fn spawn<F, T>(f: F) -> JoinHandle<T>
 where
     F: FnOnce() -> T,
     F: 'static,
     T: 'static,
 {
-    spawn_internal(f, None)
+    spawn_internal(f, None, location!())
 }
 
-fn spawn_internal<F, T>(f: F, name: Option<String>) -> JoinHandle<T>
+/// Mock implementation of `std::thread::park`.
+///
+///  Blocks unless or until the current thread's token is made available.
+///
+/// A call to `park` does not guarantee that the thread will remain parked
+/// forever, and callers should be prepared for this possibility.
+#[track_caller]
+pub fn park() {
+    rt::park(location!());
+}
+
+fn spawn_internal<F, T>(f: F, name: Option<String>, location: Location) -> JoinHandle<T>
 where
     F: FnOnce() -> T,
     F: 'static,
@@ -134,7 +159,7 @@ where
             });
 
             *result.lock().unwrap() = Some(Ok(f()));
-            notify.notify();
+            notify.notify(location);
         })
     };
 
@@ -151,6 +176,9 @@ where
 impl Builder {
     /// Generates the base configuration for spawning a thread, from which
     /// configuration methods can be chained.
+    // `std::thread::Builder` does not implement `Default`, so this type does
+    // not either, as it's a mock version of the `std` type.
+    #[allow(clippy::new_without_default)]
     pub fn new() -> Builder {
         Builder { name: None }
     }
@@ -170,20 +198,22 @@ impl Builder {
 
     /// Spawns a new thread by taking ownership of the `Builder`, and returns an
     /// `io::Result` to its `JoinHandle`.
+    #[track_caller]
     pub fn spawn<F, T>(self, f: F) -> io::Result<JoinHandle<T>>
     where
         F: FnOnce() -> T,
         F: Send + 'static,
         T: Send + 'static,
     {
-        Ok(spawn_internal(f, self.name))
+        Ok(spawn_internal(f, self.name, location!()))
     }
 }
 
 impl<T> JoinHandle<T> {
     /// Waits for the associated thread to finish.
+    #[track_caller]
     pub fn join(self) -> std::thread::Result<T> {
-        self.notify.wait();
+        self.notify.wait(location!());
         self.result.lock().unwrap().take().unwrap()
     }