+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# 0.5.4 (December 3, 2021)
+
+### Added
+
+- cell: Add `ConstPtr` and `MutPtr` RAII guards to `UnsafeCell` (#219)
+
+### Changed
+
+- Improve error message when execution state is unavailable (such as when
+  running outside of `loom::model`) (#242)
+  
+# 0.5.3 (November 23, 2021)
+
+### Added
+
+- thread: Add mock versions of `thread::park` and `Thread::unpark` (#240)
+
+### Changed
+
+- Don't attempt to clean up Mutex when threads are deadlocked (#236)
+- Update tracing-subscriber to 0.3 (#238)
+
 # 0.5.2 (October 7, 2021)
 
 ### Added